Extract collision.c across all stage overlays
