+++ config/splat.us.stcen.yaml
@@ -38,7 +38,7 @@ segments:
       - [0x1300, data]
       - [0xD42C, rodata]
       - [0xD4B0, .rodata, D600]
-      - [0xD4D8, .rodata, 11280] # EntityRedDoor
+      - [0xD4D8, .rodata, 12A3C] # EntityRedDoor
       - [0xD4F0, rodata]
       - [0xD578, .rodata, entity_relic_orb]
       - [0xD5B0, .rodata, entity_message_box]
@@ -48,7 +48,8 @@ segments:
       - [0x10E4C, c, random]
       - [0x10E7C, c, update]
       - [0x11178, c, update_stage_entities]
-      - [0x11280, c]
+      - [0x11280, c, collision]
+      - [0x12A3C, c]
       - [0x16F30, c, entity_relic_orb]
       - [0x17A00, c, entity_heart_drop]
       - [0x17B28, c, entity_message_box]
