+++ config/splat.us.stdre.yaml
@@ -50,7 +50,7 @@ segments:
       - [0x1181C, rodata]
       - [0x118D0, rodata]
       - [0x11910, .rodata, 1697C] # EntityFadeToWhite
-      - [0x1193C, .rodata, 173C4] # EntityRedDoor
+      - [0x1193C, .rodata, 18B80] # EntityRedDoor
       - [0x11954, rodata]
       - [0x119C4, .rodata, 1C7DC]
       - [0x119DC, .rodata, entity_relic_orb]
@@ -64,7 +64,8 @@ segments:
       - [0x16F90, c, random]
       - [0x16FC0, c, update]
       - [0x172BC, c, update_stage_entities]
-      - [0x173C4, c]
+      - [0x173C4, c, collision]
+      - [0x18B80, c]
       - [0x1C63C, c]
       - [0x1C7DC, c]
       - [0x1D074, c, entity_relic_orb]
