+++ config/splat.us.stnz0.yaml
@@ -89,7 +89,7 @@ segments:
       - [0x30618, .rodata, gaibon] # EntityGaibon .rodata, 33FCC
       - [0x3063C, .rodata, 36DE4] # func_801B7034
       - [0x3064C, rodata]
-      - [0x30728, .rodata, 39908] # EntityRedDoor
+      - [0x30728, .rodata, 3B0C4] # EntityRedDoor
       - [0x30740, .rodata, e_collect] # CollectGold text
       - [0x30790, .rodata, e_collect] # EntityPrizeDrop
       - [0x307B0, .rodata, e_collect] # EntityEquipItemDrop
@@ -113,7 +113,8 @@ segments:
       - [0x394D4, c, random]
       - [0x39504, c, update]
       - [0x39800, c, update_stage_entities]
-      - [0x39908, c]
+      - [0x39908, c, collision]
+      - [0x3B0C4, c, 3B0C4]
       - [0x3D848, c, e_collect]
       - [0x3F308, c, blit_char]
       - [0x3F5B8, c, entity_relic_orb]
