+++ config/splat.us.strwrp.yaml
@@ -48,7 +48,8 @@ segments:
       - [0xA168, c, random]
       - [0xA198, c, update]
       - [0xA494, c, update_stage_entities]
-      - [0xA59C, c]
+      - [0xA59C, c, collision]
+      - [0xBD58, c]
       - [0x1024C, c, entity_relic_orb]
       - [0x10D1C, c, entity_heart_drop]
       - [0x10E44, c, entity_message_box]
