+++ config/symbols.us.strwrp.txt
@@ -6,6 +6,7 @@ g_InitializeData0 = 0x8018044C;
 g_InitializeEntityData0 = 0x80180458;
 g_EInitGeneric = 0x80180470;
 g_InitDataEnt13 = 0x8018047C;
+g_eDamageDisplayInit = 0x801804A0;
 UNK_Invincibility0 = 0x80180690;
 g_testCollEnemyLookup = 0x801806BC;
 g_testCollLuckCutoff = 0x801809DC;
@@ -15,6 +16,7 @@ g_testCollElementLookup = 0x801809F0;
 g_testColliFrames = 0x80180A04;
 g_testCollPrizeTable = 0x80180A10;
 g_testCollRandTable = 0x80180A50;
+g_eDamageDisplayClut = 0x80180A6C;
 g_goldCollectTexts = 0x80180E08;
 c_GoldPrizes = 0x80180E30;
 c_HeartPrizes = 0x80180EC0;
