+++ src/st/cen/collision.c
