+++ src/st/mad/10544.c
@@ -1,64 +1,101 @@
-#include "np3.h"
+#include "mad.h"
 
-#include "../hit_detection.h"
+#include "../st_private.h"
 
-#include "../entity_damage_display.h"
+#include "../create_entity_from_layout.h"
+#include "../create_entity_in_range.h"
+#include "../find_entity_horizontal.h"
+#include "../create_entities_right.h"
 
-#include "../libstage.h"
+/*
+ * n.b.! This is different from every other stage's `CreateEntitiesToTheLeft`.
+ * It will at most create 1 entity to the left and then exit with the horizontal
+ * array pointer updated to the next element in the sequence.
+ */
+void CreateEntitiesToTheLeft(s16 posX) {
+    u8 flag;
+    s32 expected = 0;
+
+    if (posX < 0) {
+        posX = 0;
+    }
+
+    if (g_LayoutObjPosHorizontal == LAYOUT_OBJ_POSITION_FORWARD) {
+        FindFirstEntityToTheLeft(posX - g_ScrollDeltaX);
+        g_LayoutObjPosHorizontal = LAYOUT_OBJ_POSITION_BACKWARD;
+    }
+
+    if (g_LayoutObjHorizontal[LAYOUT_OBJ_POS_X] == LAYOUT_OBJ_START ||
+        g_LayoutObjHorizontal[LAYOUT_OBJ_POS_X] < posX) {
+        return;
+    }
+
+    flag = (g_LayoutObjHorizontal[LAYOUT_OBJ_SLOT] >> 8) + 0xff;
+    expected = 0;
+    if (flag == 0xFF ||
+        ((g_entityDestroyed[flag >> 5] & (1 << (flag & 0x1f))) == expected)) {
+        CreateEntityWhenInVerticalRange(g_LayoutObjHorizontal);
+    }
+    g_LayoutObjHorizontal -= 5;
+}
+
+#include "../find_entity_vertical.h"
+#include "../create_entities_vertical.h"
+#include "../init_room_entities.h"
+#include "../update_room_position.h"
+#include "../create_entity_from_entity.h"
 
 #include "../entity_red_door.h"
 
 #include "../entity.h"
 
-u8 func_801BC6BC(u8 frames[], Entity* self, u8 arg2) {
+u8 func_80191F24(u8 frames[], Entity* self, u8 arg2) {
     u16 animFrameStart = self->animFrameIdx * 2;
-    u8* iterEnt1 = &frames[animFrameStart];
+    u8* var_s1 = &frames[animFrameStart];
     s16 var_a1 = 0;
 
     if (self->animFrameDuration == 0) {
-        if (*iterEnt1 != 0) {
-            if (*iterEnt1 == 0xFF) {
+        if (*var_s1 != 0) {
+            if (*var_s1 == 0xFF) {
                 return 0;
             }
-            self->animFrameDuration = *iterEnt1++ + (u8)self->ext.stub[0x3F];
-            self->animCurFrame = *iterEnt1++;
+            self->animFrameDuration = *var_s1++ + (u8)self->ext.stub[0x3F];
+            self->animCurFrame = *var_s1++;
             self->animFrameIdx++;
             var_a1 = 128;
         } else {
-            iterEnt1 = frames;
+            var_s1 = frames;
             self->animFrameIdx = 0;
             self->animFrameDuration = 0;
             self->ext.stub[0x3F] = (arg2 * Random()) >> 8;
-            self->animFrameDuration = *iterEnt1++ + (u8)self->ext.stub[0x3F];
-            self->animCurFrame = *iterEnt1;
+            self->animFrameDuration = *var_s1++ + (u8)self->ext.stub[0x3F];
+            self->animCurFrame = *var_s1;
             self->animFrameIdx++;
             return 0;
         }
     }
     self->animFrameDuration--;
-    self->animCurFrame = iterEnt1[-1];
+    self->animCurFrame = var_s1[-1];
     var_a1 |= 1;
     return var_a1;
 }
 
-// Absolute distance from g_CurrentEntity to the player in the X Axis
 s32 GetDistanceToPlayerX(void) {
-    s16 yDistance = g_CurrentEntity->posX.i.hi - PLAYER.posX.i.hi;
+    s16 value = g_CurrentEntity->posX.i.hi - PLAYER.posX.i.hi;
 
-    if (yDistance < 0) {
-        yDistance = -yDistance;
+    if (value < 0) {
+        value = -value;
     }
-    return yDistance;
+    return value;
 }
 
-// Absolute distance from g_CurrentEntity to the player in the Y Axis
 s32 GetDistanceToPlayerY(void) {
-    s32 yDistance = g_CurrentEntity->posY.i.hi - PLAYER.posY.i.hi;
+    s32 value = g_CurrentEntity->posY.i.hi - PLAYER.posY.i.hi;
 
-    if (yDistance < 0) {
-        yDistance = -yDistance;
+    if (value < 0) {
+        value = -value;
     }
-    return yDistance;
+    return value;
 }
 
 #include "../get_side_to_player.h"
@@ -74,7 +111,23 @@ void FallEntity(void) {
     }
 }
 
-s32 func_801BC8E4(u16* sensors) {
+u8 func_8019214C(void) {
+    u8 unkState;
+    Entity* entity;
+
+    MoveEntity();
+    FallEntity();
+
+    entity = g_CurrentEntity;
+
+    if (unkState & 1) {
+        entity->velocityY = 0;
+    }
+
+    return unkState;
+}
+
+s32 func_80192190(u16* sensors) {
     Collider col;
     Collider colBack;
     s16 x;
@@ -127,7 +180,9 @@ s32 func_801BC8E4(u16* sensors) {
     return 0;
 }
 
-s32 func_801BCB5C(s16* posX) {
+INCLUDE_ASM("asm/us/st/mad/nonmatchings/10544", func_80192408);
+
+s32 func_80192618(s16* posX) {
     Collider collider;
     s16 temp2;
     s16 temp4;
@@ -190,39 +245,40 @@ Entity* AllocEntity(Entity* start, Entity* end) {
             DestroyEntity(current);
             return current;
         }
+
         current++;
     }
     return NULL;
 }
 
-s32 func_801BCDA4(u8 arg0, s16 arg1) { return D_80181A50[arg0] * arg1; }
+s32 func_80192860(u8 arg0, s16 arg1) { return D_801809EC[arg0 & 0xFF] * arg1; }
 
-s16 func_801BCDD0(u8 arg0) { return D_80181A50[arg0]; }
+s16 func_8019288C(u8 arg0) { return D_801809EC[arg0]; }
 
-void func_801BCDEC(s32 arg0, s16 arg1) {
-    g_CurrentEntity->velocityX = func_801BCDA4(arg0, arg1);
-    g_CurrentEntity->velocityY = func_801BCDA4(arg0 - 0x40, arg1);
+void func_801928A8(s32 arg0, s16 arg1) {
+    g_CurrentEntity->velocityX = func_80192860(arg0, arg1);
+    g_CurrentEntity->velocityY = func_80192860(arg0 - 0x40, arg1);
 }
 
-u8 func_801BCE58(s16 x, s16 y) { return (ratan2(y, x) >> 4) + 0x40; }
+u8 func_80192914(s16 x, s16 y) { return (ratan2(y, x) >> 4) + 0x40; }
 
-u8 func_801BCE90(Entity* a, Entity* b) {
+u8 func_8019294C(Entity* a, Entity* b) {
     s32 diffX = (u16)b->posX.i.hi - (u16)a->posX.i.hi;
     s32 diffY = (u16)b->posY.i.hi - (u16)a->posY.i.hi;
-    return func_801BCE58(diffX, diffY);
+    return func_80192914(diffX, diffY);
 }
 
-u8 func_801BCED8(s32 x, s32 y) {
+u8 func_80192994(s32 x, s32 y) {
     s32 diffX = x - (u16)g_CurrentEntity->posX.i.hi;
     s32 diffY = y - (u16)g_CurrentEntity->posY.i.hi;
-    return func_801BCE58(diffX, diffY);
+    return func_80192914(diffX, diffY);
 }
 
 #include "../adjust_value_within_threshold.h"
 
 #include "../unk_entity_func0.h"
 
-u16 func_801BD004(s16 x, s16 y) { return ratan2(y, x); }
+u16 func_80192AC0(s16 x, s16 y) { return ratan2(y, x); }
 
 u16 GetAngleBetweenEntities(Entity* a, Entity* b) {
     s32 diffX = b->posX.i.hi - a->posX.i.hi;
@@ -230,13 +286,13 @@ u16 GetAngleBetweenEntities(Entity* a, Entity* b) {
     return ratan2(diffY, diffX);
 }
 
-u16 func_801BD06C(s32 x, s32 y) {
+u16 func_80192B28(s32 x, s32 y) {
     s16 diffX = x - (u16)g_CurrentEntity->posX.i.hi;
     s16 diffY = y - (u16)g_CurrentEntity->posY.i.hi;
     return ratan2(diffY, diffX);
 }
 
-u16 GetNormalizedAngle(u16 arg0, s16 arg1, s16 arg2) {
+u16 GetNormalizedAngle(u16 arg0, u16 arg1, u16 arg2) {
     u16 var_v0 = arg1;
     u16 temp_a2 = arg2 - arg1;
     u16 var_v0_2;
@@ -260,39 +316,68 @@ u16 GetNormalizedAngle(u16 arg0, s16 arg1, s16 arg2) {
     return arg2;
 }
 
-void SetStep(u8 step) {
-    g_CurrentEntity->step = step;
+void func_80192BD0(s32 arg0) {
+    g_CurrentEntity->step = (s16)(arg0 & 0xFF);
     g_CurrentEntity->step_s = 0;
     g_CurrentEntity->animFrameIdx = 0;
     g_CurrentEntity->animFrameDuration = 0;
 }
 
-void SetSubStep(u8 step_s) {
-    g_CurrentEntity->step_s = step_s;
+void func_80192BF0(s32 arg0) {
+    g_CurrentEntity->step_s = (s16)(arg0 & 0xFF);
     g_CurrentEntity->animFrameIdx = 0;
     g_CurrentEntity->animFrameDuration = 0;
 }
 
-void EntityExplosionSpawn(u16 arg0, u16 sfxId) {
-    if (sfxId != 0) {
-        func_801C2598(sfxId);
-    }
+void EntityExplosionSpawn(u16 arg0, u16 arg1) {
+    Entity* entity;
 
+    if (arg1 != 0) {
+        g_api.PlaySfx(arg1);
+    }
     if (arg0 == 0xFF) {
         DestroyEntity(g_CurrentEntity);
         return;
     }
 
-    g_CurrentEntity->entityId = E_EXPLOSION;
-    g_CurrentEntity->pfnUpdate = (PfnEntityUpdate)EntityExplosion;
-    g_CurrentEntity->params = arg0;
-    g_CurrentEntity->animCurFrame = 0;
-    g_CurrentEntity->drawFlags = 0;
+    entity = g_CurrentEntity;
+    entity->drawFlags = 0;
+    entity->entityId = E_EXPLOSION;
+    entity->pfnUpdate = (PfnEntityUpdate)EntityExplosion;
+    entity->params = arg0;
+    entity->animCurFrame = 0;
     g_CurrentEntity->step = 0;
     g_CurrentEntity->step_s = 0;
 }
 
-#include "../init_entity.h"
+void InitializeEntity(u16 arg0[]) {
+    u16 enemyId;
+    EnemyDef* enemyDef;
+
+    g_CurrentEntity->animSet = *arg0++;
+    g_CurrentEntity->animCurFrame = *arg0++;
+    g_CurrentEntity->unk5A = *arg0++;
+    g_CurrentEntity->palette = *arg0++;
+
+    enemyId = *arg0++;
+    g_CurrentEntity->enemyId = enemyId;
+    enemyDef = &g_api.enemyDefs[enemyId];
+    g_CurrentEntity->hitPoints = enemyDef->hitPoints;
+    g_CurrentEntity->attack = enemyDef->attack;
+    g_CurrentEntity->attackElement = enemyDef->attackElement;
+    g_CurrentEntity->hitboxState = enemyDef->hitboxState;
+    g_CurrentEntity->hitboxWidth = enemyDef->hitboxWidth;
+    g_CurrentEntity->hitboxHeight = enemyDef->hitboxHeight;
+    g_CurrentEntity->flags = enemyDef->flags;
+    g_CurrentEntity->hitboxOffX = 0;
+    g_CurrentEntity->hitboxOffY = 0;
+    g_CurrentEntity->step_s = 0;
+    g_CurrentEntity->step++;
+    if (g_CurrentEntity->zPriority == 0) {
+        g_CurrentEntity->zPriority =
+            g_unkGraphicsStruct.g_zEntityCenter.unk - 0xC;
+    }
+}
 
 void EntityDummy(Entity* arg0) {
     if (arg0->step == 0) {
@@ -300,14 +385,15 @@ void EntityDummy(Entity* arg0) {
     }
 }
 
-s32 func_801BD308(u16* hitSensors, s16 sensorCount) {
-    s32 velocityX = g_CurrentEntity->velocityX;
+s32 func_80192DD0(u16* hitSensors, s16 sensorCount) {
     Collider collider;
+    s16 i;
+    s32 velocityX;
     u16 temp_a1;
     s16 x;
     s16 y;
-    s16 i;
 
+    velocityX = g_CurrentEntity->velocityX;
     if (velocityX != 0) {
         x = g_CurrentEntity->posX.i.hi;
         y = g_CurrentEntity->posY.i.hi;
@@ -331,19 +417,18 @@ s32 func_801BD308(u16* hitSensors, s16 sensorCount) {
     }
 }
 
-void func_801BD430(u16* hitSensors, s16 sensorCount) {
-    s32 velocityX = g_CurrentEntity->velocityX;
+void func_80192EF8(u16* hitSensors, s16 sensorCount) {
     Collider collider;
+    s16 i;
+    s32 velocityX;
     s16 x;
     s16 y;
-    s16 i;
 
+    velocityX = g_CurrentEntity->velocityX;
     if (velocityX == 0)
         return;
-
     x = g_CurrentEntity->posX.i.hi;
     y = g_CurrentEntity->posY.i.hi;
-
     for (i = 0; i < sensorCount; i++) {
         if (velocityX < 0) {
             x = x + *hitSensors++;
@@ -369,7 +454,7 @@ void func_801BD430(u16* hitSensors, s16 sensorCount) {
 
 #include "../replace_breakable_with_item_drop.h"
 
-void func_801BD984(void) {
+void func_8019344C(void) {
     s32 temp_v1;
     Entity* entity;
 
@@ -391,7 +476,7 @@ void func_801BD984(void) {
     }
 }
 
-void func_801BDA08(u16 arg0) {
+void func_801934D0(u16 arg0) {
     Collider collider;
 
     if (g_CurrentEntity->velocityX < 0) {
@@ -417,41 +502,49 @@ void func_801BDA08(u16 arg0) {
 
         if (collider.effects & EFFECT_QUICKSAND) {
             g_CurrentEntity->posY.val += FIX(0.125);
-        } else {
-            g_CurrentEntity->posY.i.hi += collider.unk18;
+            return;
         }
-    } else {
-        if (!(collider.effects & EFFECT_NOTHROUGH)) {
-            MoveEntity();
-            func_801BD984();
-        }
+
+        g_CurrentEntity->posY.i.hi =
+            (u16)g_CurrentEntity->posY.i.hi + collider.unk18;
+        return;
+    }
+
+    if (!(collider.effects & EFFECT_NOTHROUGH)) {
+        MoveEntity();
+        func_8019344C();
     }
 }
 
-#include "../collect_heart.h"
+INCLUDE_ASM("asm/us/st/mad/nonmatchings/10544", CollectHeart);
 
 #include "../collect_gold.h"
+
+void func_801937BC(void) {}
+
+void func_801937C4(void) { DestroyEntity(g_CurrentEntity); }
 
 void CollectSubweapon(u16 subWeaponIdx) {
     Entity* player = &PLAYER;
     u16 subWeapon;
 
-    g_api.PlaySfx(NA_SE_PL_IT_PICKUP);
+    g_api.PlaySfx(0x672);
     subWeapon = g_Status.subWeapon;
-    g_Status.subWeapon = D_80181D80[subWeaponIdx];
+    g_Status.subWeapon = D_80180D1C[subWeaponIdx];
 
     if (subWeapon == g_Status.subWeapon) {
         subWeapon = 1;
         g_CurrentEntity->unk6D[0] = 0x10;
     } else {
-        subWeapon = D_80181DB0[subWeapon];
+        subWeapon = D_80180D4C[subWeapon];
         g_CurrentEntity->unk6D[0] = 0x60;
     }
 
     if (subWeapon != 0) {
         g_CurrentEntity->params = subWeapon;
         g_CurrentEntity->posY.i.hi = player->posY.i.hi + 12;
-        SetStep(7);
+        g_CurrentEntity->step = 7;
+        g_CurrentEntity->step_s = 0;
         g_CurrentEntity->velocityY = FIX(-2.5);
         g_CurrentEntity->animCurFrame = 0;
         g_CurrentEntity->ext.generic.unk88.S16.unk2 = 5;
@@ -465,12 +558,23 @@ void CollectSubweapon(u16 subWeaponIdx) {
     DestroyEntity(g_CurrentEntity);
 }
 
-#include "../collect_heart_vessel.h"
+// Different from "collect_heart_vessel.h"
+void CollectHeartVessel(void) {
+    g_api.PlaySfx(0x670);
+    g_api.func_800FE044(HEART_VESSEL_INCREASE, 0x4000);
+    DestroyEntity(g_CurrentEntity);
+}
 
 void CollectLifeVessel(void) {
-    g_api.PlaySfx(NA_SE_PL_COLLECT_HEART);
-    g_api.func_800FE044(5, 0x8000);
+    g_api.PlaySfx(0x670);
+    g_api.func_800FE044(LIFE_VESSEL_INCREASE, 0x8000);
     DestroyEntity(g_CurrentEntity);
 }
 
 void DestroyCurrentEntity(void) { DestroyEntity(g_CurrentEntity); }
+
+Entity* func_801939C4(void) {
+    g_CurrentEntity->step = 3;
+    g_CurrentEntity->params = 4;
+    return g_CurrentEntity;
+}
