+++ src/st/no3/collision.c
