+++ src/st/np3/collision.c
