+++ src/st/nz0/3B0C4.c
@@ -1,7 +1,4 @@
-#include "dre.h"
-#include "../hit_detection.h"
-
-#include "../entity_damage_display.h"
+#include "nz0.h"
 
 #include "../libstage.h"
 
@@ -9,7 +6,7 @@
 
 #include "../entity.h"
 
-u8 func_8019A590(u8 frames[], Entity* self, u8 arg2) {
+u8 func_801BCAD4(u8 frames[], Entity* self, u8 arg2) {
     u16 animFrameStart = self->animFrameIdx * 2;
     u8* var_s1 = &frames[animFrameStart];
     s16 var_a1 = 0;
@@ -60,16 +57,9 @@ s32 GetDistanceToPlayerY(void) {
     return yDistance;
 }
 
-s32 GetSideToPlayer(void) {
-    s16 var_a0 = g_CurrentEntity->posX.i.hi > PLAYER.posX.i.hi;
+#include "../get_side_to_player.h"
 
-    if (g_CurrentEntity->posY.i.hi > PLAYER.posY.i.hi) {
-        var_a0 |= 2;
-    }
-    return var_a0;
-}
-
-void MoveEntity(void) {
+void MoveEntity() {
     g_CurrentEntity->posX.val += g_CurrentEntity->velocityX;
     g_CurrentEntity->posY.val += g_CurrentEntity->velocityY;
 }
@@ -80,7 +70,7 @@ void FallEntity(void) {
     }
 }
 
-s32 func_8019A7B8(u16* sensors) {
+s32 func_801BCCFC(u16* sensors) {
     Collider col;
     Collider colBack;
     s16 x;
@@ -133,7 +123,7 @@ s32 func_8019A7B8(u16* sensors) {
     return 0;
 }
 
-s32 func_8019AA30(s16* posX) {
+s32 func_801BCF74(s16* posX) {
     Collider collider;
     s16 temp2;
     s16 temp4;
@@ -202,17 +192,16 @@ Entity* AllocEntity(Entity* start, Entity* end) {
     return NULL;
 }
 
-s32 func_8019AC78(u8 arg0, s16 arg1) { return D_80180D80[arg0] * arg1; }
+s32 func_801BD1BC(u8 arg0, s16 arg1) { return D_80181978[arg0] * arg1; }
 
-extern s16 D_80180D80[];
-s16 func_8019ACA4(u8 arg0) { return D_80180D80[arg0]; }
+s16 func_801BD1E8(u8 arg0) { return D_80181978[arg0]; } // Unique
 
-void func_8019ACC0(s32 arg0, s16 arg1) {
-    g_CurrentEntity->velocityX = func_8019AC78(arg0, arg1);
-    g_CurrentEntity->velocityY = func_8019AC78(arg0 - 0x40, arg1);
+void func_801BD204(s32 arg0, s16 arg1) {
+    g_CurrentEntity->velocityX = func_801BD1BC(arg0, arg1);
+    g_CurrentEntity->velocityY = func_801BD1BC(arg0 - 0x40, arg1);
 }
 
-u8 func_8019AD2C(s16 x, s16 y) { return (ratan2(y, x) >> 4) + 0x40; }
+u8 func_801BD270(s16 x, s16 y) { return (ratan2(y, x) >> 4) + 0x40; }
 
 u8 func_8019AD64(ObjInit2* arg0, ObjInit2* arg1) {
     u16 x, y;
@@ -220,23 +209,23 @@ u8 func_8019AD64(ObjInit2* arg0, ObjInit2* arg1) {
     x = arg1->zPriority - arg0->zPriority;
     y = arg1->palette - arg0->palette;
 
-    return func_8019AD2C(x, y);
+    return func_801BD270(x, y);
 }
 
-u8 func_8019ADAC(s16 arg0, s16 arg1) {
+u8 func_801BD2F0(s16 arg0, s16 arg1) {
     s16 x, y;
 
     x = arg0 - g_CurrentEntity->posX.i.hi;
     y = arg1 - g_CurrentEntity->posY.i.hi;
 
-    return func_8019AD2C(x, y);
+    return func_801BD270(x, y);
 }
 
 #include "../adjust_value_within_threshold.h"
 
 #include "../unk_entity_func0.h"
 
-u16 func_8019AED8(s16 arg0, s16 arg1) { return ratan2(arg1, arg0); }
+u16 func_801BD41C(s16 x, s16 y) { return ratan2(y, x); }
 
 u16 GetAngleBetweenEntities(Entity* a, Entity* b) {
     s32 diffX = b->posX.i.hi - a->posX.i.hi;
@@ -244,12 +233,13 @@ u16 GetAngleBetweenEntities(Entity* a, Entity* b) {
     return ratan2(diffY, diffX);
 }
 
-u16 func_8019AF40(s32 x, s32 y) {
+u16 func_801BD484(s32 x, s32 y) {
     s16 diffX = x - (u16)g_CurrentEntity->posX.i.hi;
     s16 diffY = y - (u16)g_CurrentEntity->posY.i.hi;
     return ratan2(diffY, diffX);
 }
 
+// TODO(sestren): Reconcile this GetNormalizedAngle with the one in 3E30C.c
 u16 GetNormalizedAngle(u16 arg0, s16 arg1, s16 arg2) {
     u16 var_v0 = arg1;
     u16 temp_a2 = arg2 - arg1;
@@ -275,29 +265,21 @@ u16 GetNormalizedAngle(u16 arg0, s16 arg1, s16 arg2) {
 }
 
 void SetStep(u8 step) {
-    Entity* entity;
-
-    entity = g_CurrentEntity;
-    entity->step = step;
-    entity->step_s = 0;
-    entity->animFrameIdx = 0;
-    entity->animFrameDuration = 0;
+    g_CurrentEntity->step = step;
+    g_CurrentEntity->step_s = 0;
+    g_CurrentEntity->animFrameIdx = 0;
+    g_CurrentEntity->animFrameDuration = 0;
 }
 
 void SetSubStep(u8 step_s) {
-    Entity* entity;
-
-    entity = g_CurrentEntity;
-    entity->step_s = step_s;
-    entity->animFrameIdx = 0;
-    entity->animFrameDuration = 0;
+    g_CurrentEntity->step_s = step_s;
+    g_CurrentEntity->animFrameIdx = 0;
+    g_CurrentEntity->animFrameDuration = 0;
 }
 
-void EntityExplosionSpawn(u16 arg0, u16 arg1) {
-    Entity* entity;
-
-    if (arg1 != 0) {
-        func_801A046C(arg1);
+void EntityExplosionSpawn(u16 arg0, u16 sfxId) {
+    if (sfxId != 0) {
+        func_801C29B0(sfxId);
     }
 
     if (arg0 == 0xFF) {
@@ -305,25 +287,51 @@ void EntityExplosionSpawn(u16 arg0, u16 arg1) {
         return;
     }
 
-    entity = g_CurrentEntity;
-    entity->drawFlags = 0;
-    entity->entityId = E_EXPLOSION;
-    entity->pfnUpdate = (PfnEntityUpdate)EntityExplosion;
-    entity->params = arg0;
-    entity->animCurFrame = 0;
+    g_CurrentEntity->entityId = E_EXPLOSION;
+    g_CurrentEntity->pfnUpdate = (PfnEntityUpdate)EntityExplosion;
+    g_CurrentEntity->params = arg0;
+    g_CurrentEntity->animCurFrame = 0;
+    g_CurrentEntity->drawFlags = 0;
     g_CurrentEntity->step = 0;
     g_CurrentEntity->step_s = 0;
 }
 
-#include "../init_entity.h"
+void InitializeEntity(u16 arg0[]) {
+    u16 enemyId;
+    EnemyDef* enemyDef;
 
-void func_8019B1B4(Entity* arg0) {
+    g_CurrentEntity->animSet = *arg0++;
+    g_CurrentEntity->animCurFrame = *arg0++;
+    g_CurrentEntity->unk5A = *arg0++;
+    g_CurrentEntity->palette = *arg0++;
+
+    enemyId = *arg0++;
+    g_CurrentEntity->enemyId = enemyId;
+    enemyDef = &g_api.enemyDefs[enemyId];
+    g_CurrentEntity->hitPoints = enemyDef->hitPoints;
+    g_CurrentEntity->attack = enemyDef->attack;
+    g_CurrentEntity->attackElement = enemyDef->attackElement;
+    g_CurrentEntity->hitboxState = enemyDef->hitboxState;
+    g_CurrentEntity->hitboxWidth = enemyDef->hitboxWidth;
+    g_CurrentEntity->hitboxHeight = enemyDef->hitboxHeight;
+    g_CurrentEntity->flags = enemyDef->flags;
+    g_CurrentEntity->hitboxOffX = 0;
+    g_CurrentEntity->hitboxOffY = 0;
+    g_CurrentEntity->step_s = 0;
+    g_CurrentEntity->step++;
+    if (g_CurrentEntity->zPriority == 0) {
+        g_CurrentEntity->zPriority =
+            g_unkGraphicsStruct.g_zEntityCenter.unk - 0xC;
+    }
+}
+
+void EntityDummy(Entity* arg0) {
     if (arg0->step == 0) {
         arg0->step++;
     }
 }
 
-s32 func_8019B1DC(u16* hitSensors, s16 sensorCount) {
+s32 func_801BD720(u16* hitSensors, s16 sensorCount) {
     Collider collider;
     s16 i;
     s32 velocityX;
@@ -355,29 +363,31 @@ s32 func_8019B1DC(u16* hitSensors, s16 sensorCount) {
     }
 }
 
-void func_8019B304(u16* hitSensors, s16 sensorCount) {
+void CheckFieldCollision(u16* hitSensors, s16 sensorCount) {
+    s32 velocityX = g_CurrentEntity->velocityX;
     Collider collider;
+    s16 x, y;
     s16 i;
-    s32 velocityX;
-    s16 x;
-    s16 y;
 
-    velocityX = g_CurrentEntity->velocityX;
-    if (velocityX == 0)
+    if (velocityX == 0) {
         return;
+    }
+
     x = g_CurrentEntity->posX.i.hi;
     y = g_CurrentEntity->posY.i.hi;
+
     for (i = 0; i < sensorCount; i++) {
         if (velocityX < 0) {
-            x = x + *hitSensors++;
+            x += *hitSensors++;
         } else {
-            x = x - *hitSensors++;
+            x -= *hitSensors++;
         }
-
         y += *hitSensors++;
+
         g_api.CheckCollision(x, y, &collider, 0);
-        if (collider.effects & EFFECT_UNK_0002 &&
-            (!(collider.effects & EFFECT_UNK_8000) || i != 0)) {
+
+        if ((collider.effects & EFFECT_UNK_0002) &&
+            (!(collider.effects & EFFECT_UNK_8000) || (i != 0))) {
             if (velocityX < 0) {
                 g_CurrentEntity->posX.i.hi += collider.unk1C;
             } else {
@@ -391,125 +401,3 @@ void func_8019B304(u16* hitSensors, s16 sensorCount) {
 #include "../get_player_collision_with.h"
 
 #include "../replace_breakable_with_item_drop.h"
-
-void func_8019B858(void) {
-    s32 temp_v1;
-    Entity* entity = g_CurrentEntity;
-
-    if (entity->velocityY >= 0) {
-        temp_v1 =
-            entity->ext.generic.unk84.unk + entity->ext.generic.unk88.S16.unk0;
-        entity->ext.generic.unk84.unk = temp_v1;
-        entity->velocityX = temp_v1;
-
-        if ((temp_v1 == 0x10000) || (temp_v1 == -0x10000)) {
-            entity->ext.generic.unk88.U16.unk0 =
-                -entity->ext.generic.unk88.U16.unk0;
-        }
-        entity = g_CurrentEntity;
-    }
-
-    if (entity->velocityY < FIX(0.25)) {
-        entity->velocityY += FIX(0.125);
-    }
-}
-
-void func_8019B8DC(u16 arg0) {
-    Collider collider;
-
-    if (g_CurrentEntity->velocityX < 0) {
-        g_api.CheckCollision(g_CurrentEntity->posX.i.hi,
-                             g_CurrentEntity->posY.i.hi - 7, &collider, 0);
-        if (collider.effects & EFFECT_NOTHROUGH) {
-            g_CurrentEntity->velocityY = 0;
-        }
-    }
-
-    g_api.CheckCollision(g_CurrentEntity->posX.i.hi,
-                         g_CurrentEntity->posY.i.hi + 7, &collider, 0);
-
-    if (arg0) {
-        if (!(collider.effects & EFFECT_NOTHROUGH)) {
-            MoveEntity();
-            FallEntity();
-            return;
-        }
-
-        g_CurrentEntity->velocityX = 0;
-        g_CurrentEntity->velocityY = 0;
-
-        if (collider.effects & EFFECT_QUICKSAND) {
-            g_CurrentEntity->posY.val += FIX(0.125);
-            return;
-        }
-
-        g_CurrentEntity->posY.i.hi += collider.unk18;
-        return;
-    }
-
-    if (!(collider.effects & EFFECT_NOTHROUGH)) {
-        MoveEntity();
-        func_8019B858();
-    }
-}
-
-void func_8019BA38(u16 arg0) {
-    s32* hearts;
-
-    g_api.PlaySfx(NA_SE_PL_COLLECT_HEART);
-    hearts = &g_Status.hearts;
-    *hearts += c_HeartPrizes[arg0];
-
-    if (g_Status.heartsMax < *hearts) {
-        *hearts = g_Status.heartsMax;
-    }
-
-    DestroyEntity(g_CurrentEntity);
-}
-
-#include "../collect_gold.h"
-
-void CollectSubweapon(u16 subWeaponIdx) {
-    Entity* player = &PLAYER;
-    u16 subWeapon;
-
-    g_api.PlaySfx(NA_SE_PL_IT_PICKUP);
-    subWeapon = g_Status.subWeapon;
-    g_Status.subWeapon = D_801810B0[subWeaponIdx];
-
-    if (subWeapon == g_Status.subWeapon) {
-        subWeapon = 1;
-        g_CurrentEntity->unk6D[0] = 0x10;
-    } else {
-        subWeapon = D_801810E0[subWeapon];
-        g_CurrentEntity->unk6D[0] = 0x60;
-    }
-
-    if (subWeapon != 0) {
-        g_CurrentEntity->params = subWeapon;
-        g_CurrentEntity->posY.i.hi = player->posY.i.hi + 12;
-        SetStep(7);
-        g_CurrentEntity->velocityY = FIX(-2.5);
-        g_CurrentEntity->animCurFrame = 0;
-        g_CurrentEntity->ext.generic.unk88.S16.unk2 = 5;
-        if (player->facingLeft != 1) {
-            g_CurrentEntity->velocityX = FIX(-2);
-            return;
-        }
-        g_CurrentEntity->velocityX = FIX(2);
-        return;
-    }
-    DestroyEntity(g_CurrentEntity);
-}
-
-#include "../collect_heart_vessel.h"
-
-void CollectLifeVessel(void) {
-    g_api.PlaySfx(NA_SE_PL_COLLECT_HEART);
-    g_api.func_800FE044(LIFE_VESSEL_INCREASE, 0x8000);
-    DestroyEntity(g_CurrentEntity);
-}
-
-void DestroyCurrentEntity(void) { DestroyEntity(g_CurrentEntity); }
-
-INCLUDE_ASM("st/dre/nonmatchings/173C4", EntityPrizeDrop);
