+++ src/st/rwrp/BD58.c
@@ -1,14 +1,10 @@
 #include "rwrp.h"
 
-#include "../hit_detection.h"
-
-INCLUDE_ASM("st/rwrp/nonmatchings/A59C", EntityDamageDisplay);
-
 #include "../libstage.h"
 
 #include "../entity_is_near_player.h"
 
-INCLUDE_ASM("st/rwrp/nonmatchings/A59C", func_8018C948);
+INCLUDE_ASM("st/rwrp/nonmatchings/BD58", func_8018C948);
 
 #include "../entity.h"
 
@@ -73,7 +69,7 @@ void FallEntity(void) {
         g_CurrentEntity->velocityY += FALL_GRAVITY;
     }
 }
-INCLUDE_ASM("st/rwrp/nonmatchings/A59C", func_8018D990);
+INCLUDE_ASM("st/rwrp/nonmatchings/BD58", func_8018D990);
 
 s32 func_8018DC08(s16* posX) {
     Collider collider;
@@ -243,7 +239,7 @@ void EntityExplosionSpawn(u16 arg0, u16 sfxId) {
     g_CurrentEntity->step_s = 0;
 }
 
-INCLUDE_ASM("st/rwrp/nonmatchings/A59C", InitializeEntity);
+INCLUDE_ASM("st/rwrp/nonmatchings/BD58", InitializeEntity);
 
 void EntityDummy(Entity* entity) {
     if (entity->step == 0) {
@@ -251,9 +247,9 @@ void EntityDummy(Entity* entity) {
     }
 }
 
-INCLUDE_ASM("st/rwrp/nonmatchings/A59C", func_8018E3B4);
+INCLUDE_ASM("st/rwrp/nonmatchings/BD58", func_8018E3B4);
 
-INCLUDE_ASM("st/rwrp/nonmatchings/A59C", func_8018E4DC);
+INCLUDE_ASM("st/rwrp/nonmatchings/BD58", func_8018E4DC);
 
 #include "../get_player_collision_with.h"
 
@@ -281,7 +277,7 @@ void func_8018EA30(void) {
     }
 }
 
-INCLUDE_ASM("st/rwrp/nonmatchings/A59C", func_8018EAB4);
+INCLUDE_ASM("st/rwrp/nonmatchings/BD58", func_8018EAB4);
 
 extern s8 c_HeartPrizes[];
 
@@ -332,7 +328,7 @@ void CollectLifeVessel(void) {
 
 void DestroyCurrentEntity(void) { DestroyEntity(g_CurrentEntity); }
 
-INCLUDE_ASM("st/rwrp/nonmatchings/A59C", EntityPrizeDrop);
+INCLUDE_ASM("st/rwrp/nonmatchings/BD58", EntityPrizeDrop);
 
 #define MISSING_ANIMATE_ENTITY_PROTOTYPE
 #include "../entity_explosion.h"
@@ -340,6 +336,6 @@ INCLUDE_ASM("st/rwrp/nonmatchings/A59C", EntityPrizeDrop);
 
 #include "../blink_item.h"
 
-INCLUDE_ASM("st/rwrp/nonmatchings/A59C", EntityEquipItemDrop);
+INCLUDE_ASM("st/rwrp/nonmatchings/BD58", EntityEquipItemDrop);
 
 #include "../blit_char.h"
